调整ZoneLocator继承结构，AbstractZoneLocator.doGet方法移动到HttpUtils
