+++ microsphere-multiactive-spring-cloud/src/main/java/io/microsphere/multiple/active/zone/aws/Ec2AvailabilityZoneEndpointZoneLocator.java
@@ -1,9 +1,14 @@
 package io.microsphere.multiple.active.zone.aws;
 
 import io.microsphere.multiple.active.zone.AbstractZoneLocator;
+import io.microsphere.multiple.active.zone.HttpUtils;
 import io.microsphere.multiple.active.zone.ZoneLocator;
+import org.springframework.context.EnvironmentAware;
 import org.springframework.core.env.Environment;
 import org.springframework.util.StringUtils;
+
+import static io.microsphere.multiple.active.zone.ZoneConstants.DEFAULT_TIMEOUT;
+import static io.microsphere.multiple.active.zone.ZoneConstants.LOCATOR_TIMEOUT_PROPERTY_NAME;
 
 /**
  * Amazon EC2 Availability Zone Endpoint {@link ZoneLocator}
@@ -13,13 +18,15 @@ import org.springframework.util.StringUtils;
  * @see ZoneLocator
  * @since 1.0.0
  */
-public class Ec2AvailabilityZoneEndpointZoneLocator extends AbstractZoneLocator {
+public class Ec2AvailabilityZoneEndpointZoneLocator extends AbstractZoneLocator implements EnvironmentAware {
 
     public static final String AVAILABILITY_ZONE_ENDPOINT_URI_PROPERTY_NAME = "EC2_AVAILABILITY_ZONE_ENDPOINT_URI";
 
     public static final String DEFAULT_AVAILABILITY_ZONE_ENDPOINT_URI = "http://169.254.169.254/latest/meta-data/placement/availability-zone";
 
     public static final int DEFAULT_ORDER = 15;
+
+    private int timeout = DEFAULT_TIMEOUT;
 
     public Ec2AvailabilityZoneEndpointZoneLocator() {
         super(DEFAULT_ORDER);
@@ -36,7 +43,7 @@ public class Ec2AvailabilityZoneEndpointZoneLocator extends AbstractZoneLocator 
         String zone = null;
         if (StringUtils.hasText(uri)) {
             try {
-                zone = doGet(uri);
+                zone = HttpUtils.doGet(uri, timeout);
                 logger.info("The zone ['{}'] was located from the EC2 Availability Zone Endpoint[URI : '{}' , property name: '{}']", zone, uri,
                         AVAILABILITY_ZONE_ENDPOINT_URI_PROPERTY_NAME);
             } catch (Throwable e) {
@@ -51,5 +58,10 @@ public class Ec2AvailabilityZoneEndpointZoneLocator extends AbstractZoneLocator 
 
     private String getAvailabilityZoneEndpointURI(Environment environment) {
         return environment.getProperty(AVAILABILITY_ZONE_ENDPOINT_URI_PROPERTY_NAME, DEFAULT_AVAILABILITY_ZONE_ENDPOINT_URI);
+    }
+
+    @Override
+    public void setEnvironment(Environment environment) {
+        this.timeout = environment.getProperty(LOCATOR_TIMEOUT_PROPERTY_NAME, int.class, DEFAULT_TIMEOUT);
     }
 }
